Add new `mime` package
