+++ package.json
@@ -76,6 +76,7 @@
 		"build:docs:media-recording-actions": "docgen packages/media-recording/src/store/actions.ts --output packages/media-recording/README.md --to-token --use-token \"Autogenerated actions|src/store/actions.ts\"",
 		"build:docs:media-recording-selectors": "docgen packages/media-recording/src/store/selectors.ts --output packages/media-recording/README.md --to-token --use-token \"Autogenerated selectors|src/store/selectors.ts\"",
 		"build:docs:media-utils": "docgen packages/media-utils/src/index.ts --output packages/media-utils/README.md --to-token --use-token \"Autogenerated API docs\"",
+		"build:docs:mime": "docgen packages/mime/src/index.ts --output packages/mime/README.md --to-token --use-token \"Autogenerated API docs\"",
 		"build:docs:pdf": "docgen packages/pdf/src/index.ts --output packages/pdf/README.md --to-token --use-token \"Autogenerated API docs\"",
 		"build:docs:subtitles": "docgen packages/subtitles/src/index.ts --output packages/subtitles/README.md --to-token --use-token \"Autogenerated API docs\"",
 		"build:docs:upload-media-actions": "docgen packages/upload-media/src/store/actions.ts --output packages/upload-media/README.md --to-token --use-token \"Autogenerated actions|src/store/actions.ts\"",
