+++ packages/editor/package.json
@@ -36,6 +36,7 @@
 		"@mexp/interface": "../interface",
 		"@mexp/media-recording": "../media-recording",
 		"@mexp/media-utils": "../media-utils",
+		"@mexp/mime": "../mime",
 		"@mexp/upload-media": "../upload-media",
 		"@wordpress/blob": "^4.0.0",
 		"@wordpress/block-editor": "^13.0.0",
