+++ packages/editor/tsconfig.json
@@ -7,7 +7,7 @@
 	"references": [
 		{ "path": "../interface" },
 		{ "path": "../media-recording" },
-		{ "path": "../media-utils" },
+		{ "path": "../mime" },
 		{ "path": "../upload-media" }
 	],
 	"include": [ "src/**/*" ]
