+++ packages/ffmpeg/package.json
@@ -34,6 +34,7 @@
 		"@ffmpeg/core": "^0.11.0",
 		"@ffmpeg/ffmpeg": "^0.11.6",
 		"@mexp/media-utils": "../media-utils",
+		"@mexp/mime": "../mime",
 		"uuid": "^10.0.0"
 	}
 }
