+++ packages/media-recording/package.json
@@ -31,7 +31,7 @@
 	},
 	"dependencies": {
 		"@mediapipe/selfie_segmentation": "^0.1.1675465747",
-		"@mexp/media-utils": "../media-utils",
+		"@mexp/mime": "../mime",
 		"@wordpress/blob": "^4.0.0",
 		"@wordpress/data": "^10.0.0",
 		"@wordpress/date": "^5.0.0",
