+++ packages/media-utils/package.json
@@ -29,8 +29,5 @@
 		".": {
 			"default": "./src/index.ts"
 		}
-	},
-	"dependencies": {
-		"mime": "^3.0.0"
 	}
 }
