+++ packages/mime/src/test/getExtensionFromMimeType.ts
