+++ packages/mime/src/test/getMediaTypeFromMimeType.ts
