+++ packages/mime/src/test/getMimeTypeFromExtension.ts
