+++ packages/mime/tsconfig.json
@@ -4,6 +4,5 @@
 		"rootDir": "src",
 		"declarationDir": "dist-types"
 	},
-	"references": [ { "path": "../media-utils" }, { "path": "../mime" } ],
 	"include": [ "src/**/*" ]
 }
