+++ packages/upload-media/package.json
@@ -34,6 +34,7 @@
 		"@mexp/heif": "../heif",
 		"@mexp/log": "../log",
 		"@mexp/media-utils": "../media-utils",
+		"@mexp/mime": "../mime",
 		"@mexp/pdf": "../pdf",
 		"@mexp/subtitles": "../subtitles",
 		"@mexp/vips": "../vips",
