+++ packages/upload-media/tsconfig.json
@@ -9,6 +9,7 @@
 		{ "path": "../heif" },
 		{ "path": "../log" },
 		{ "path": "../media-utils" },
+		{ "path": "../mime" },
 		{ "path": "../pdf" },
 		{ "path": "../subtitles" },
 		{ "path": "../vips" }
