+++ packages/vips/package.json
@@ -31,7 +31,7 @@
 		}
 	},
 	"dependencies": {
-		"@mexp/media-utils": "../media-utils",
+		"@mexp/mime": "../mime",
 		"wasm-vips": "^0.0.9"
 	}
 }
