+++ tsconfig.json
@@ -22,6 +22,9 @@
 			"path": "packages/media-utils"
 		},
 		{
+			"path": "packages/mime"
+		},
+		{
 			"path": "packages/pdf"
 		},
 		{
